Rename test, remove visualization stuff
