+++ multibody/parsing/BUILD.bazel
@@ -386,7 +386,7 @@ drake_cc_googletest(
 )
 
 drake_cc_googletest(
-    name = "drake_package_uri_test",
+    name = "drake_manifest_resolution_test",
     data = [
         ":test_models",
     ],
