+++ multibody/parsing/test/drake_manifest_resolution_test/cube_visual.urdf
